tasks/itinerary-task/requirements.txt
# Itinerary Task Worker dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
python-consul==1.1.0

